# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Portal DPC PERADI SAI Medan MUSCAB II
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Stats {
  calon: number;
  peserta: number;
}

export interface OkResponse {
  ok: boolean;
}

export type AdminSessionRole = typeof AdminSessionRole[keyof typeof AdminSessionRole];


export const AdminSessionRole = {
  full_access: 'full_access',
  registrasi: 'registrasi',
  keuangan: 'keuangan',
  spectator: 'spectator',
} as const;

export interface AdminSession {
  username: string;
  role: AdminSessionRole;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminUser {
  id: string;
  username: string;
  role: AdminSessionRole;
  createdAt?: string;
}

export interface AdminUserInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  role: AdminSessionRole;
}

export interface AdminUserUpdate {
  /** @minLength 6 */
  password?: string;
  role?: AdminSessionRole;
}

export type PesertaStatusNia = typeof PesertaStatusNia[keyof typeof PesertaStatusNia];


export const PesertaStatusNia = {
  pending: 'pending',
  aktif: 'aktif',
  tidak_aktif: 'tidak_aktif',
} as const;

export type PesertaStatusBayar = typeof PesertaStatusBayar[keyof typeof PesertaStatusBayar];


export const PesertaStatusBayar = {
  menunggu: 'menunggu',
  terverifikasi: 'terverifikasi',
  ditolak: 'ditolak',
} as const;

export interface Peserta {
  id: string;
  nama: string;
  nia: string;
  alamat: string;
  hp: string;
  email: string;
  statusNia: PesertaStatusNia;
  statusBayar: PesertaStatusBayar;
  hadir: boolean;
  /** @nullable */
  selfiePath?: string | null;
  /** @nullable */
  buktiPath?: string | null;
  /** @nullable */
  fotoKtaPath?: string | null;
  /** @nullable */
  catatanAdmin?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface PesertaInput {
  /** @minLength 1 */
  nama: string;
  /** @minLength 1 */
  nia: string;
  /** @minLength 1 */
  alamat: string;
  /** @minLength 1 */
  hp: string;
  email: string;
  /** @nullable */
  selfiePath?: string | null;
  /** @nullable */
  buktiPath?: string | null;
  /** @nullable */
  fotoKtaPath?: string | null;
}

export type PesertaUpdateStatusNia = typeof PesertaUpdateStatusNia[keyof typeof PesertaUpdateStatusNia];


export const PesertaUpdateStatusNia = {
  pending: 'pending',
  aktif: 'aktif',
  tidak_aktif: 'tidak_aktif',
} as const;

export type PesertaUpdateStatusBayar = typeof PesertaUpdateStatusBayar[keyof typeof PesertaUpdateStatusBayar];


export const PesertaUpdateStatusBayar = {
  menunggu: 'menunggu',
  terverifikasi: 'terverifikasi',
  ditolak: 'ditolak',
} as const;

export interface PesertaUpdate {
  statusNia?: PesertaUpdateStatusNia;
  statusBayar?: PesertaUpdateStatusBayar;
  hadir?: boolean;
  /** @nullable */
  catatanAdmin?: string | null;
}

export type PesertaPublikStatusNia = typeof PesertaPublikStatusNia[keyof typeof PesertaPublikStatusNia];


export const PesertaPublikStatusNia = {
  pending: 'pending',
  aktif: 'aktif',
  tidak_aktif: 'tidak_aktif',
} as const;

export type PesertaPublikStatusBayar = typeof PesertaPublikStatusBayar[keyof typeof PesertaPublikStatusBayar];


export const PesertaPublikStatusBayar = {
  menunggu: 'menunggu',
  terverifikasi: 'terverifikasi',
  ditolak: 'ditolak',
} as const;

export interface PesertaPublik {
  id: string;
  nama: string;
  nia: string;
  statusNia: PesertaPublikStatusNia;
  statusBayar: PesertaPublikStatusBayar;
  createdAt: string;
}

export interface SuaraData {
  ketua: number[];
  sekjend: number[];
}

export type SuaraUpdateInputPosisi = typeof SuaraUpdateInputPosisi[keyof typeof SuaraUpdateInputPosisi];


export const SuaraUpdateInputPosisi = {
  ketua: 'ketua',
  sekjend: 'sekjend',
} as const;

export interface SuaraUpdateInput {
  posisi: SuaraUpdateInputPosisi;
  /**
     * @minimum 0
     * @maximum 1
     */
  calonIndex: number;
  /** @minimum 0 */
  jumlahSuara: number;
}

export interface MemberSession {
  nama: string;
  email: string;
}

export interface MemberRegisterInput {
  /** @minLength 1 */
  nama: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface MemberLoginInput {
  email: string;
  password: string;
}

export interface ForumTopic {
  id: string;
  judul: string;
  isi: string;
  namaPenulis: string;
  jumlahBalasan: number;
  createdAt: string;
}

export interface ForumTopicInput {
  /** @minLength 3 */
  judul: string;
  /** @minLength 1 */
  isi: string;
}

export interface ForumReply {
  id: string;
  isi: string;
  namaPenulis: string;
  createdAt: string;
}

export interface ForumReplyInput {
  /** @minLength 1 */
  isi: string;
}

export type ForumTopicDetail = ForumTopic & {
  replies: ForumReply[];
};

export type ListAdminPesertaParams = {
search?: string;
};

export type CekPesertaParams = {
search?: string;
};