# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Portal DPC PERADI SAI Medan MUSCAB II
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Login admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Kredensial salah
...
  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Info admin yang sedang login
      responses:
        "200":
          description: Data admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Belum login
  /admin/peserta:
    get:
      operationId: listAdminPeserta
      tags: [admin]
      summary: Daftar semua peserta (admin)
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Daftar peserta
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Peserta"
  /admin/peserta/{id}:
    patch:
      operationId: updateAdminPeserta
      tags: [admin]
      summary: Update data peserta (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PesertaUpdate"
      responses:
        "200":
          description: Data peserta diupdate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peserta"
        "404":
          description: Peserta tidak ditemukan
    delete:
      operationId: deleteAdminPeserta
      tags: [admin]
      summary: Hapus peserta (full_access only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Peserta dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: Daftar semua admin user (full_access only)
      responses:
        "200":
          description: Daftar admin users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: createAdminUser
      tags: [admin]
      summary: Buat admin user baru (full_access only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Admin user dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update admin user (full_access only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Admin user diupdate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Hapus admin user (full_access only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Admin user dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /forum/topics:
    get:
      operationId: listForumTopics
      tags: [forum]
      summary: Daftar topik forum diskusi
      responses:
        "200":
          description: Daftar topik
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForumTopic"
    post:
      operationId: createForumTopic
      tags: [forum]
      summary: Buat topik diskusi baru (harus login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForumTopicInput"
      responses:
        "201":
          description: Topik dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumTopic"
        "401":
          description: Belum login
  /forum/topics/{id}:
    get:
      operationId: getForumTopic
      tags: [forum]
      summary: Detail topik & balasan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Detail topik
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumTopicDetail"
        "404":
          description: Topik tidak ditemukan
  /forum/topics/{id}/replies:
    post:
      operationId: createForumReply
      tags: [forum]
      summary: Kirim balasan pada topik (harus login)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForumReplyInput"
      responses:
        "201":
          description: Balasan dikirim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumReply"
        "401":
          description: Belum login
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    AdminSessionRole:
      type: string
      enum: [full_access, registrasi, keuangan, spectator]
    AdminSession:
      type: object
      properties:
        username:
          type: string
        role:
          $ref: "#/components/schemas/AdminSessionRole"
      required:
        - username
        - role
    AdminLoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AdminUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        role:
          $ref: "#/components/schemas/AdminSessionRole"
        createdAt:
          type: string
      required:
        - id
        - username
        - role
    AdminUserInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        role:
          $ref: "#/components/schemas/AdminSessionRole"
      required:
        - username
        - password
        - role
    AdminUserUpdate:
      type: object
      properties:
        password:
          type: string
          minLength: 6
        role:
          $ref: "#/components/schemas/AdminSessionRole"
    Peserta:
      type: object
      properties:
        id:
          type: string
        nama:
          type: string
        nia:
          type: string
        alamat:
          type: string
        hp:
          type: string
        email:
          type: string
        statusNia:
          type: string
          enum: [pending, aktif, tidak_aktif]
        statusBayar:
          type: string
          enum: [menunggu, terverifikasi, ditolak]
        hadir:
          type: boolean
        selfiePath:
          type: ["string", "null"]
        buktiPath:
          type: ["string", "null"]
        fotoKtaPath:
          type: ["string", "null"]
        catatanAdmin:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - nama
        - nia
        - alamat
        - hp
        - email
        - statusNia
        - statusBayar
        - hadir
        - createdAt
        - updatedAt
...
    PesertaUpdate:
      type: object
      properties:
        statusNia:
          type: string
          enum: [pending, aktif, tidak_aktif]
        statusBayar:
          type: string
          enum: [menunggu, terverifikasi, ditolak]
        hadir:
          type: boolean
        catatanAdmin:
          type: ["string", "null"]
...
    ForumTopic:
      type: object
      properties:
        id:
          type: string
        judul:
          type: string
        isi:
          type: string
        namaPenulis:
          type: string
        jumlahBalasan:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - judul
        - isi
        - namaPenulis
        - jumlahBalasan
        - createdAt
    ForumTopicInput:
      type: object
      properties:
        judul:
          type: string
          minLength: 3
        isi:
          type: string
          minLength: 1
      required:
        - judul
        - isi
    ForumReply:
      type: object
      properties:
        id:
          type: string
        isi:
          type: string
        namaPenulis:
          type: string
        createdAt:
          type: string
      required:
        - id
        - isi
        - namaPenulis
        - createdAt
    ForumReplyInput:
      type: object
      properties:
        isi:
          type: string
          minLength: 1
      required:
        - isi
    ForumTopicDetail:
      allOf:
        - $ref: "#/components/schemas/ForumTopic"
        - type: object
          properties:
            replies:
              type: array
              items:
                $ref: "#/components/schemas/ForumReply"
          required:
            - replies