# lib/api-zod/src/generated/types/adminUserInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Portal DPC PERADI SAI Medan MUSCAB II
 * OpenAPI spec version: 0.1.0
 */
import type { AdminSessionRole } from './adminSessionRole';

export interface AdminUserInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  role: AdminSessionRole;
}