# lib/api-zod/src/generated/types/forumReplyInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Portal DPC PERADI SAI Medan MUSCAB II
 * OpenAPI spec version: 0.1.0
 */

export interface ForumReplyInput {
  /** @minLength 1 */
  isi: string;
}