# lib/api-zod/src/generated/types/forumTopicDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Portal DPC PERADI SAI Medan MUSCAB II
 * OpenAPI spec version: 0.1.0
 */
import type { ForumReply } from './forumReply';
import type { ForumTopic } from './forumTopic';

export type ForumTopicDetail = ForumTopic & {
  replies: ForumReply[];
};