# lib/api-zod/src/generated/types/pesertaUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Portal DPC PERADI SAI Medan MUSCAB II
 * OpenAPI spec version: 0.1.0
 */
import type { PesertaUpdateStatusBayar } from './pesertaUpdateStatusBayar';
import type { PesertaUpdateStatusNia } from './pesertaUpdateStatusNia';

export interface PesertaUpdate {
  statusNia?: PesertaUpdateStatusNia;
  statusBayar?: PesertaUpdateStatusBayar;
  hadir?: boolean;
  /** @nullable */
  catatanAdmin?: string | null;
}